+++ R/zzz.R
@@ -1,6 +1,6 @@
 .onAttach <- function(...) {
   packageStartupMessage(paste0(
     "haldensify v", utils::packageDescription("haldensify")$Version,
-    ": Conditional Density Estimation with the Highly Adaptive Lasso"
+    ": ", utils::packageDescription("haldensify")$Title
   ))
 }