+++ NEWS.md
@@ -1,7 +1,7 @@
 # haldensify 0.0.5
 
 * First CRAN release.
-
+* Documentation updates
 
 # haldensify 0.0.3